更新主題和插件還有 gitignore 規則和 nginx 設定
Blocksy Theme
Blocksy Companion
Elementor
ElementsKit Lite
EmbedPress
Essential Addons for Elementor
Limit Login Attempts Reloaded
LiteSpeed Cache
Rank Math SEO
Sucuri Security - Auditing, Malware Scanner and Hardening

+++ wordpress/wp-content/plugins/blocksy-companion/framework/dashboard.php
@@ -195,6 +195,10 @@ class Dashboard {
 				return;
 			}
 
+			if (is_network_admin()) {
+				return;
+			}
+
 			if (intval(get_option('blc_activation_redirect', false)) === wp_get_current_user()->ID) {
 				delete_option('blc_activation_redirect');
 				exit(wp_redirect(admin_url('admin.php?page=ct-dashboard')));
@@ -295,6 +299,14 @@ class Dashboard {
 
 	public function is_dashboard_page() {
 		global $pagenow;
+
+		if (is_network_admin()) {
+			$is_ct_settings =
+				// 'themes.php' === $pagenow &&
+				isset( $_GET['page'] ) && 'blocksy-companion' === $_GET['page'];
+
+			return $is_ct_settings;
+		}
 
 		$is_ct_settings =
 			// 'themes.php' === $pagenow &&
